requirements.txt
Flask==2.0.2
Pillow==8.4.0
numpy==1.21.5
keras==2.6.0
